Add git-pushtip
It's used in a workflow in which each developer has a tip/USER/
namespace. It can be used together with git-send-branch to setup a
review through mainling list, pushing to the remote server for
convenience.

It will try to autoconfigure itself, but is known to work only on
gitolite as remote. If using anything else, you can configure it by
setting:

	git config --add pushtip.remote <remote-name>
	git config --add pushtip.branchprefix tip/<USER>