arch-installer: fail if gdisk failed
If gdisk fails, for example because it's not installer, the script was
not failing earlier but rather generating an invalid image.

This makes sure the script is now failing in that case (and also if
gdisk fails for other reasons). We need to check the return code of the
piped command, hence the usage of PIPESTATUS env var.