arch-installer: make sure we have a current package database
I had this fail on me claiming that there were no package databases for
core and extra. This surprised me, but forcing a "pacman -Syy" prior to
running "pacman -Sg base" makes sure this doesn't happen.

This always allows us to fail gracefully if there is a network problem
(e.g. if we are behind a proxy).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>