arch-installer: don't write new files in /dev/
A common mistake is to try to write the image into /dev/sd* but for some
reason the device block is not there (USB not being properly
identified). Without this check, a new file would be created hiding the
mistake. So, assume that if the passed file is in /dev/ it must exist.