git-fetchpr: fix ambiguity while checking out new branch
Depending on the branch name, git may be confused if it's a path or a
branch name the argument to rev-parse:

++ git rev-parse review/pr-2722-pxf-battery
fatal: ambiguous argument 'review/pr-2722-pxf-battery': unknown revision
or path not in the working tree.
Use '--' to separate paths from revisions, like this:
'git <command> [<revision>...] -- [<file>...]'

Use a '--' in the end to fix it.