git-applypr: replace non-existent amend-trailer
I had a local git-amend-trailer that was doing:

	GIT_EDITOR="git interpret-trailers --if-exists=addIfDifferent --where=end --trailer \"$t\" --in-place" \
			git commit --amend >/dev/null

To apply the trailer by looping through the args. However git-commit
also has a --trailer option and we can override the config to git to
pass what are our desired arguments. With that, we can simply embed
the logic in git-applypr itself.

Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>