kdeploy: Do not require remote to have sudo without password
It's much more common to have sudo with password. Make it the default,
in which case we go through the following hops:

1) rsync to a local tmpdir (either created or passed in the command line)
2) rsync to a remote tmpdir, default to /tmp/kdeploy
3) ssh to remote to send script to execute
4) ssh to remote to finish execution, with sudo, possibly asking for
   password:
	- rsync from remote tmpdir to /
	- initrd
	- bootloader
	- reboot

When --remote-tmpdir= or --remote-tmpdir=no is passed, one rsync is
avoided, but we still have same amount of logins:

1) rsync to a local tmpdir (either created or passed in the command line)
2) rsync to remote /
3) ssh to remote to send script to execute
4) ssh to remote to finish execution, with sudo, possibly asking for
   password:
	- initrd
	- bootloader
	- reboot

(3) and (4) is needed because we can't use ssh's stdin to send the
script, otherwise a pty isn't created so sudo can't ask for password.
In future we may move (3) to be done together with (2).

Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>